// data/project-index/extracted-project/src/app/modules/partner/business/models/business-payments.model.ts
export class recordMultipleData {
    TotalRows: number | null;
    CustomerId: string | null;
    ProductId: string | null;
    Units: string | null;
    UnitOfMeasure: string | null;
    Description: string | null;
    CostToPartner: string | null;
    CurrencyCode: string | null;
    UsageDate: string | null;
    ValidationStatus: string | null;
    ValidationError: any;
    ImportStatus: string | null;
    ImportError: any;
    ValidationProcessedCount: number | null;
    ValidationSuccessCount: number | null;
    ValidationErrorCount: number | null;
    ImportProcessedCount: number | null;
    ImportSuccessCount: number | null;
    ImportErrorCount: number | null;
    ProductName: string | null;
    CustomerName: string | null;
    CurrencySymbol: string | null;
    CurrencyDecimalPlaces: string | null;
    CurrencyDecimalSeperator: string | null;
    CurrencyThousandSeperator: string | null;
}